Propagate background and effects styles from the root to the RenderView. https://bugs.webkit.org/show_bug.cgi?id=245032 <rdar://99782177>
Reviewed by NOBODY (OOPS!).

Propagates the background and effect styles from the root element up to the
RenderView, and marks the old styles so that they return an empty used value.

Sets m_hasInitializedStyle=true when setting style on a RenderElement, otherwise
this never gets set to true for RenderView (which doesn't initialize style
during creation). This is needed to get future style changes to correctly report
a diff.

Adds a ChildChangeInvalidation case for removing the body, this makes sure we
recompute the RenderView's style in case it had propagated the old body's
background.

* Source/WTF/Scripts/Preferences/UnifiedWebPreferences.yaml:
* Source/WebCore/rendering/RenderElement.cpp:
(WebCore::RenderElement::setStyle):
* Source/WebCore/style/ChildChangeInvalidation.cpp:
(WebCore::Style::ChildChangeInvalidation::invalidateAfterChange):
* Source/WebCore/style/StyleAdjuster.cpp:
(WebCore::Style::Adjuster::propagateToDocumentElementAndInitialContainingBlock):