`fillText` after `transferControlToScreen` for canvas doesn't draw. https://bugs.webkit.org/show_bug.cgi?id=290042 <rdar://problem/147925254>
Reviewed by NOBODY (OOPS!).

PlaceholderRenderingContextSource::setPlaceholderBuffer can be called before the
delegate is provided (either due to a race, or due to the <canvas> element for
which this is a placeholder not being in a Document).

In this case, we need to hold onto the ImageBuffer and copy it to the delegate
when/if it becomes available. Make sure we only hold references to the
ImageBuffer on a single thread, and use WTF_GUARDED_BY_CAPABILITY to enforce
this.

* Source/WebCore/html/canvas/PlaceholderRenderingContext.cpp:
(WebCore::PlaceholderRenderingContextSource::setPlaceholderBuffer):
(WebCore::PlaceholderRenderingContextSource::setContentsToLayer):
(WebCore::PlaceholderRenderingContext::setPlaceholderBuffer):
* Source/WebCore/html/canvas/PlaceholderRenderingContext.h:
(WebCore::PlaceholderRenderingContextSource::WTF_GUARDED_BY_CAPABILITY):